feat: score page

+++ lib/app/core/shared/utilities/number/thousand_formatter.dart
@@ -1,6 +1,6 @@
 import 'package:intl/intl.dart';
 
-String thousandFormatter(int number) {
+String thousandFormatter(double number) {
   NumberFormat numberFormat = NumberFormat.decimalPattern("pt");
 
   return numberFormat.format(number);
